backend/src/config.py
from dotenv import load_dotenv
import os

# Cargar variables de entorno desde el archivo .env
load_dotenv()

# Azure AI Search
AZURE_SEARCH_ENDPOINT = os.getenv("AZURE_SEARCH_ENDPOINT")
AZURE_SEARCH_API_KEY = os.getenv("AZURE_SEARCH_API_KEY")
AZURE_SEARCH_INDEX_NAME = os.getenv("AZURE_SEARCH_INDEX_NAME")

# Azure OpenAI
AZURE_OPENAI_ENDPOINT = os.getenv("AZURE_OPENAI_ENDPOINT")
AZURE_OPENAI_API_KEY = os.getenv("AZURE_OPENAI_API_KEY")
AZURE_OPENAI_CHAT_DEPLOYMENT = os.getenv("AZURE_OPENAI_CHAT_DEPLOYMENT")
AZURE_OPENAI_EMBEDDING_DEPLOYMENT = os.getenv("AZURE_OPENAI_EMBEDDING_DEPLOYMENT")

# Modelo más pequeño para clasificación (opcional)
AZURE_OPENAI_CLASSIFICATION_DEPLOYMENT = os.getenv("AZURE_OPENAI_CLASSIFICATION_DEPLOYMENT", "gpt-4.1-mini")

# Cosmos DB
COSMOS_ENDPOINT = os.getenv("COSMOS_ENDPOINT")
COSMOS_KEY = os.getenv("COSMOS_KEY")
COSMOS_DATABASE_NAME = os.getenv("COSMOS_DATABASE_NAME", "chat_database")
COSMOS_CONTAINER_NAME = os.getenv("COSMOS_CONTAINER_NAME", "chat_history")

# Flag para habilitar/deshabilitar Cosmos DB
USE_COSMOS_DB = os.getenv("USE_COSMOS_DB", "false").lower() == "true"
